generate-recipe/requirements.txt
functions-framework==3.*
google-generativeai
google-cloud-aiplatform
google-cloud-translate